@@ -86,7 +86,7 @@
 								</requireMavenVersion>
 								<requireJavaVersion>
 									<!-- enforce =1.8 so that we compile against the right JRE -->
-									<version>[1.8,)</version>
+									<version>[17,)</version>
 								</requireJavaVersion>
 							</rules>
 						</configuration>
@@ -101,8 +101,8 @@
 					<artifactId>maven-compiler-plugin</artifactId>
 					<version>3.8.1</version>
 					<configuration>
-						<source>1.8</source>
-						<target>1.8</target>
+						<source>17</source>
+						<target>17</target>
 					</configuration>
 				</plugin>
 				<plugin>
