@@ -11,7 +11,7 @@ public class SettingsException extends SAMLException {
 	public static final int PUBLIC_CERT_FILE_NOT_FOUND = 5;    
 	public static final int PRIVATE_KEY_FILE_NOT_FOUND = 6;
 
-	private int errorCode;
+	private final int errorCode;
 
 	public SettingsException(String message, int errorCode) {
 		super(message);
