@@ -55,7 +55,7 @@ public class ValidationError extends SAMLException {
 	public static final int MISSING_ENCRYPTED_ELEMENT = 48;
 	public static final int INVALID_ISSUE_INSTANT_FORMAT = 49;
 
-	private int errorCode;
+	private final int errorCode;
 
 	public ValidationError(String message, int errorCode) {
 		super(message);
